dom/media/tests/mochitest/test_a_noOp.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1264772
-->
<head>
  <title>Test for Bug 1264772</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1264772">Mozilla Bug 1264772</a>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 1264772 **/
// The WebRTC tests seem to have more problems with intermittents (at
// least on Android) if they run first in a test run. This is a dummy test
// to ensure that the browser is ready prior to running any actual WebRTC
// tests.
//
// Note: mochitests are run in alphabetical order, so it is not sufficient
//       for this test to appear first in the manifest.
ok(true, "test passed");

</script>
</pre>
</body>
